+++ src/parser/tests/parse_errors.rs
@@ -1,3 +1,9 @@
 mod parse_error_containers_tests;
+mod parse_error_function_tests;
+mod parse_error_initializer_tests;
 mod parse_error_literals_tests;
+mod parse_error_messages_test;
+mod parse_error_program_tests;
 mod parse_error_statements_tests;
+mod parse_error_type_tests;
+mod parse_error_variable_tests;
