+++ src/parser/tests/parse_errors/parse_error_variable_tests.rs
@@ -1,0 +1,82 @@
+use crate::parser::{parse, tests::lex};
+
+#[test]
+fn empty_global_vars_can_be_parsed() {
+    let lexer = lex("VAR_GLOBAL END_VAR");
+    let result = parse(lexer).unwrap().0;
+
+    let vars = &result.global_vars[0]; //globar_vars
+    let ast_string = format!("{:#?}", vars);
+    let expected_ast = r#"VariableBlock {
+    variables: [],
+    variable_block_type: Global,
+}"#;
+    assert_eq!(ast_string, expected_ast)
+}
+
+#[test]
+fn global_vars_can_be_parsed() {
+    let lexer = lex("VAR_GLOBAL x : INT; y : BOOL; END_VAR");
+    let result = parse(lexer).unwrap().0;
+
+    let vars = &result.global_vars[0]; //globar_vars
+    let ast_string = format!("{:#?}", vars);
+    let expected_ast = r#"VariableBlock {
+    variables: [
+        Variable {
+            name: "x",
+            data_type: DataTypeReference {
+                referenced_type: "INT",
+            },
+        },
+        Variable {
+            name: "y",
+            data_type: DataTypeReference {
+                referenced_type: "BOOL",
+            },
+        },
+    ],
+    variable_block_type: Global,
+}"#;
+    assert_eq!(ast_string, expected_ast)
+}
+
+#[test]
+fn two_global_vars_can_be_parsed() {
+    let lexer = lex("VAR_GLOBAL a: INT; END_VAR VAR_GLOBAL x : INT; y : BOOL; END_VAR");
+    let result = parse(lexer).unwrap().0;
+
+    let vars = &result.global_vars; //global_vars
+    let ast_string = format!("{:#?}", vars);
+    let expected_ast = r#"[
+    VariableBlock {
+        variables: [
+            Variable {
+                name: "a",
+                data_type: DataTypeReference {
+                    referenced_type: "INT",
+                },
+            },
+        ],
+        variable_block_type: Global,
+    },
+    VariableBlock {
+        variables: [
+            Variable {
+                name: "x",
+                data_type: DataTypeReference {
+                    referenced_type: "INT",
+                },
+            },
+            Variable {
+                name: "y",
+                data_type: DataTypeReference {
+                    referenced_type: "BOOL",
+                },
+            },
+        ],
+        variable_block_type: Global,
+    },
+]"#;
+    assert_eq!(ast_string, expected_ast)
+}
